Changed structure, fixed Modal
1. Added Modal folder
2. Replace Modal file and change component file type to .jsx
3. Added Modal.styled and index.js files
4. Added useEffect to useModal script for prevent body scrolling
