+++ src/components/Modal/index.js
@@ -1,0 +1,1 @@
+export {Modal} from './Modal';
